+++ aioswaggerpy_test/aioswtest.py
@@ -21,9 +21,8 @@ async def listen_ws():
                                   media='sound:hello-world')
                 await ari.channels.continueInDialplan(channelId=channelId)
 
-ari = AsyncSwaggerClient(None, None, loop=loop, http_client=http_client)
-loop.run_until_complete(ari.do_init("http://192.168.254.60:8088/ari/api-docs/resources.json"))
-
-
+ari = AsyncSwaggerClient(None, None, loop=loop, http_client=http_client,
+                         url="http://192.168.254.60:8088/ari/api-docs/resources.json")
+loop.run_until_complete(ari.async_init())
 loop.run_until_complete(listen_ws())
-ari.close()
+ari.async_close()