+++ CHANGELOG.rst
@@ -1,6 +1,6 @@
 0.2.0 (2013-10-28)
 ------------------
-- Add close() methods to client and http_client.
+- Add async_close() methods to client and http_client.
 
 0.1.0 (2013-10-24)
 ------------------